Added TaskService, UserService, and some shit that i don't remember

+++ DataAccessLayer/DataAccessLayerCore/Entities/DailyTask.cs
@@ -3,8 +3,8 @@ using DataAccessLayerCore.Enum;
 namespace DataAccessLayerCore.Entities{
 
     public class DailyTask : BaseId{
-        public int TaskId { get; set; }
-        public int UserId { get; set; }      
+        public long TaskId { get; set; }
+        public long UserId { get; set; }      
         public string Title { get; set; } = null!;
         public string? Description { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
